requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pandas>=1.5.0
matplotlib>=3.6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
